Add brightness controls.
Refactor argument parsing to allow different kinds of scalars (not just speed).
Define brightness as a scalar type and expose a feature to set the brightness dynamically. This feature is completely necessary to operate this tool on some systems, where the brightness is reset to zero as soon as GRUB is loaded.
Does some other minor tweaking that adds little value but is generally unobjectionable.