Update README.md according to #82
If the keyboard's lights default to a brightness of 0 when booting into Linux, the user might mistakenly think their commands are not correctly sent or interpreted. This confusion is especially likely to occur if the brightness defaults to a higher value when booting into e.g. Windows. I've updated the README to instruct users to try the brightness command in case their keyboard remains dark. 

Closes #82